% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-dependencies.R
\name{add_dsfr_deps}
\alias{add_dsfr_deps}
\title{dsfr dependencies utils}
\usage{
add_dsfr_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches dsfr.  dependencies to the given tag
}
