% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.dsfr-dependencies.R
\name{add_shiny.dsfr_deps}
\alias{add_shiny.dsfr_deps}
\title{shiny.dsfr dependencies utils}
\usage{
add_shiny.dsfr_deps(tag)
}
\arguments{
\item{tag}{Element to attach the dependencies.}
}
\description{
This function attaches shiny.dsfr dependencies to the given tag
}
