% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dollar_to_br}
\alias{dollar_to_br}
\title{Substitute dollar sign by br tag}
\usage{
dollar_to_br(string)
}
\arguments{
\item{string}{string}
}
\value{
tagList
}
\description{
Substitute dollar sign by br tag
}
