% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-accordiongroup.R
\name{dsfrAccordionsGroup}
\alias{dsfrAccordionsGroup}
\title{Accordions Group}
\usage{
dsfrAccordionsGroup(..., id = NULL, expand = c("none", "first"))
}
\arguments{
\item{expand}{}
}
\value{
Accordions group
}
\description{
Accordions Group
}
