% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-footer.R
\name{dsfrFooter}
\alias{dsfrFooter}
\title{Ajoute la section du footer}
\usage{
dsfrFooter(
  marianne_title = "République$française",
  marianne_href = "/",
  marianne_href_title = "Retour à l'accueil",
  description = NULL,
  content_links = list(list(href = "https://legifrance.gouv.fr", title =
    "legifrance.gouv.fr"), list(href = "https://gouvernement.fr", title =
    "gouvernement.fr"), list(href = "https://service-public.fr", title =
    "service-public.fr"), list(href = "https://data.gouv.fr", title = "data.gouv.fr")),
  bottom_links = NULL
)
}
\arguments{
\item{marianne_title}{Le titre utilisé dans le bloc marianne du footer}

\item{marianne_href}{Le lien de redirection (accueil par defaut)}

\item{marianne_href_title}{Le titre utilisé dans l'info bulle}

\item{content_links}{Une liste de lien vers d'autres contenus
list(
list(href= "https://legifrance.gouv.fr", title = "legifrance.gouv.fr"),
list(href= "https://gouvernement.fr", title = "gouvernement.fr"),
list(href= "https://service-public.fr", title = "service-public.fr"),
list(href= "https://data.gouv.fr", title = "data.gouv.fr"))}

\item{bottom_links}{Une liste de liens divers}
}
\value{
le footer
}
\description{
Ajoute la section du footer
}
\examples{
 if (interactive()) {
 dsfrFooter(
 marianne_title = "République$française",
marianne_href = "/",
 marianne_href_title = "Retour à l'accueil",
 description = NULL,
 content_links = list(
   list(href = "https://legifrance.gouv.fr", title ="legifrance.gouv.fr"),
   list(href = "https://gouvernement.fr", title ="gouvernement.fr"),
   list(href = "https://service-public.fr", title ="service-public.fr"),
   list(href = "https://data.gouv.fr", title = "data.gouv.fr")),
 bottom_links = list(
 list(href = "#", title = "Plan du site"),
 list(href = "#", title = "Accessibilité conforme"),
 list(href = "#", title = "Mentions légales"),
 list(href = "#", title = "Données personnelles"))
)
}

}
