% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-header.R
\name{dsfrHeader}
\alias{dsfrHeader}
\title{Ajoute une entête}
\usage{
dsfrHeader(
  ...,
  marianne_title = "République$française",
  service_title = NULL,
  service_baseline = NULL,
  service_href = "/",
  service_href_title = service_title,
  logo_src = NULL,
  logo_alt = "logo",
  logo_style_width = "width:3.5rem;",
  enlarge_link = TRUE
)
}
\arguments{
\item{...}{Menu de navigation}

\item{marianne_title}{Le nom du ministère ou de la prefecture}

\item{service_title}{Le nom du service}

\item{service_baseline}{Précision sur le service ou l'application}

\item{service_href}{Le lien de redirection (Accueil par défaut)}

\item{service_href_title}{Le titre du lien affiché en infobulle}

\item{logo_src}{Chemin vers le logo}

\item{logo_alt}{Texte alternatif si le logo ne s'affiche pas}

\item{logo_style_width}{Taille du logo}

\item{enlarge_link}{Elargit le lien au bloc (TRUE par defaut)}
}
\value{
l'entête
}
\description{
Ajoute une entête
}
\examples{
if (interactive()) {
  dsfrHeader(
    marianne_title = "République$française",
    service_title = "Mon service",
    service_baseline = "Une information sur le service"
  )
}

}
