% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-navsidebarpanel.R
\name{dsfrNavSidebarPanel}
\alias{dsfrNavSidebarPanel}
\title{Ajoute un menu latéral de navigation}
\usage{
dsfrNavSidebarPanel(
  ...,
  position = c("left", "right"),
  fluid = FALSE,
  gutters = FALSE,
  horizontal_align = "center",
  vertical_align = "top",
  spacing_class = "fr-my-4w",
  width = 3,
  hide_panel = TRUE,
  id = NULL,
  title = NULL,
  title_burger = "Dans cette rubrique"
)
}
\arguments{
\item{...}{dsfrTabPanel...}

\item{position}{La position du panneau de navigation : "left" ou "right"}

\item{fluid}{Pour utiliser les marges externes ou non : TRUE (sans marges) FALSE (avec marge par défaut)}

\item{gutters}{Pour utiliser les gouttières : TRUE (avec) FALSE (sans par défaut)}

\item{horizontal_align}{Pour aligner le contenu horizontalement ("center" par défaut, "left", "right")}

\item{vertical_align}{Pour aligner le contenu verticalement ("middle", "top", "bottom")}

\item{spacing_class}{Pour ajuster l'espacement :  "fr-my-4w" (par défaut)}

\item{width}{La largeur du panneau, entre 1 et 12 : 3 (par défaut)}

\item{hide_panel}{Pour activer le javascript afin de masquer les dsfrTabPanel associé}

\item{id}{Identifiant du panneau (facultatif mais conseillé)}

\item{title}{Le titre du panneau de navigation}

\item{title_burger}{Le titre du menu burger du panneau : "Dans cette rubrique" (par défaut)}
}
\value{
Renvoie le contenu avec un menu latéral de navigation
}
\description{
Ajoute un menu latéral de navigation
}
