% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-page.R
\name{dsfrPage}
\alias{dsfrPage}
\title{Create a page with dsfr layout}
\usage{
dsfrPage(
  ...,
  title = "",
  lang = "fr",
  theme = c("system", "light", "dark"),
  suppress_bootstrap = TRUE
)
}
\arguments{
\item{...}{Elements to include within the page}

\item{title}{The browser title}

\item{lang}{The lang (default "fr")}

\item{theme}{The theme ("system", "dark", "light")}
}
\description{
Create a page with dsfr layout
}
