% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-input.R
\name{dsfrSelectInput}
\alias{dsfrSelectInput}
\title{dsfrSelectInput}
\usage{
dsfrSelectInput(
  inputId,
  label,
  choices,
  selected = NULL,
  hint = NULL,
  placeHolder = "S"
)
}
\arguments{
\item{selected}{}
}
\description{
dsfrSelectInput
}
