% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-sidebarlayout.R
\name{dsfrSidebarLayout}
\alias{dsfrSidebarLayout}
\title{SidebarLayout}
\usage{
dsfrSidebarLayout(
  sidebarPanel,
  mainPanel,
  position = c("left", "right"),
  fluid = FALSE,
  gutters = FALSE,
  horizontal_align = "center",
  vertical_align = "top",
  spacing_class = "fr-my-4w"
)
}
\arguments{
\item{spacing_class}{}
}
\description{
SidebarLayout
}
