% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-sidebarpanel.R
\name{dsfrSidebarPanel}
\alias{dsfrSidebarPanel}
\title{dsfrSidebarPanel}
\usage{
dsfrSidebarPanel(
  ...,
  width = 4,
  style = htmltools::css(`background-color` = "#f6f6f6", border = "solid", `bord-width`
    = "1px", `border-color` = "#e5e5e5")
)
}
\arguments{
\item{width}{}
}
\value{
sideparnel
}
\description{
dsfrSidebarPanel
}
