% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-tabpanel.R
\name{dsfrTabPanel}
\alias{dsfrTabPanel}
\title{Tabpanel}
\usage{
dsfrTabPanel(..., id = NULL, title = NULL, role = "tabpanel")
}
\arguments{
\item{...}{tagList}

\item{id}{id}

\item{title}{the title is keep}

\item{role}{tabpanel, navbarpanel, navbarsidepanel}
}
\value{
tabpanel
}
\description{
Tabpanel
}
