% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-tabsetpanel.R
\name{dsfrTabsetPanel}
\alias{dsfrTabsetPanel}
\title{dsfrTabsetPanel}
\usage{
dsfrTabsetPanel(..., id = NULL)
}
\arguments{
\item{...}{dsfrTabPanel(role = tabpanel)}

\item{id}{the tabsetpanel id}
}
\value{
tabsetPanel
}
\description{
dsfrTabsetPanel
}
\examples{

if (interactive()) {
dsfrTabsetPanel(id = "mon_panel1",
                dsfrTabPanel(h4("Mon contenu 1"), id = "tab1", title = "Onglet 1", role = "tabpanel"),
                dsfrTabPanel(h4("Mon contenu 2"), id = "tab2", title = "Onglet 2", role = "tabpanel")
                )
}

}
