% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-input.R
\name{dsfrToggleSwitch}
\alias{dsfrToggleSwitch}
\title{Toggle switch}
\usage{
dsfrToggleSwitch(
  inputId,
  value = FALSE,
  label = NULL,
  hint = NULL,
  position_label_left = FALSE
)
}
\arguments{
\item{inputId}{The id of the input object.}

\item{value}{checked \code{TRUE} or unchecked \code{FALSE}.}

\item{label}{The label to set for the input object.}

\item{hint}{The hint to set for the input object.}

\item{position_label_left}{\code{FALSE} (default) or \code{TRUE} to position the label to the left}
}
\description{
Toggle switch
}
