% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-footer.R
\name{footer_list_wrapper}
\alias{footer_list_wrapper}
\title{Function to generate the footer wrapper list}
\usage{
footer_list_wrapper(footer_class = "content")
}
\arguments{
\item{footer_class}{c("content", "bottom")}
}
\value{
ul wrapper
}
\description{
Function to generate the footer wrapper list
}
