% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-footer.R
\name{generate_footer_link}
\alias{generate_footer_link}
\title{Function to generate the footer link}
\usage{
generate_footer_link(link = NULL, footer_class = "content")
}
\arguments{
\item{link}{a list with href and title}

\item{footer_class}{c("content", "bottom")}
}
\value{
tag
}
\description{
Function to generate the footer link
}
