% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-footer.R
\name{generate_footer_links}
\alias{generate_footer_links}
\title{Generate all block footer links}
\usage{
generate_footer_links(links = NULL, footer_class = NULL)
}
\arguments{
\item{footer_class}{}
}
\description{
Generate all block footer links
}
