% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsfr-input.R
\name{updatedDsfrToggleSwitch}
\alias{updatedDsfrToggleSwitch}
\title{Change the value of toggle switch on the client}
\usage{
updatedDsfrToggleSwitch(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  hint = NULL,
  value = NULL
)
}
\arguments{
\item{session}{The session object passed to function given to shinyServer.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{hint}{The hint to set for the input object.}

\item{value}{The value to set for the input object.}
}
\description{
Change the value of toggle switch on the client
}
